- Implemented the CountDownTimer. I have extended the 'Countdown Timer' class under a inner class called CountDownTimer in the CountingScreen class.
I have linked a  start and a stop button for testing purposes for the
timer. However I will change this in the future to begin the timer once
the user has begun their first count.

- Fixed up some unnecessary code in the CountingScreen which can be
easily done with 1 line of code instead of multiple lines of code (The
one where you set the current selected object).

- Created a xml file which serves the purpose of having a rectangular
backround. This can be accessed by using the following code in xml,
@drawable/backgroundrectangle.

- Added a rectangular background for the currently selected object as
well as the Total count.